Increase default maximum index size for non-regtest chains (#448)
The ordinal index, stored in <ORD_DATA_DIR>/index.redb, is a redb
database, which grows dynamically. However, the maximum size that the
database can grow to must be declared up-front, to avoid expensive
region allocation operations on insert. We've been using 10 MiB as the
default index size, which is too small for anything but testing.

This commit changes the default maximum index size to 1 TiB on mainnet,
testnet, and signet, and leaves it at 10 MiB on regtest. 1 TiB is more
than needed, as even the mainnet database is only ~50 GiB, but
allocating a database is fast enough, so it doesn't hurt to over
provision and not have to think about it. For testing though, which is
done with `--chain=regtest`, we want database allocation is fast, and 10
MiB is more than enough.